bootc-image-builder: merge test_progress into test_opts
We were running these two test files together since they're both
relatively small and quick.  Merge the two files so that our dynamic
gitlab CI pipeline generator doesn't need to have any special
conditions.
